Project Setup/adk-setup/requirements.txt
google-adk
google-cloud-aiplatform[adk,agent_engines]
python-dotenv
